Broom methods for tidy cuminc objects

Broom methods for tidy cuminc objects

# S3 method for tidycuminc
tidy(x, conf.int = FALSE, conf.level = 0.95, times = NULL, ...)

# S3 method for tidycuminc
glance(x, ...)

Arguments

x
object of class 'tidycuminc'

conf.int
Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level
The confidence level to use for the confidence interval if conf.int = TRUE. Must be strictly greater than 0 and less than 1. Defaults to 0.95, which corresponds to a 95 percent confidence interval.

times
Numeric vector of times to obtain risk estimates at

...
not used

Value

a tibble

tidy() confidence intervals

The confidence intervals in tidy() use the recommended method in Competing Risks: A Practical Perspective by Melania Pintilie.

x^exp(-z * se / (x * log(x))), x^exp(z * se / (x * log(x)))

See also

Examples

cuminc <- cuminc(Surv(ttdeath, death_cr) ~ trt, trial)

tidy(cuminc)
#> # A tibble: 444 × 5
#>    outcome           strata  time estimate std.error
#>    <chr>             <chr>  <dbl>    <dbl>     <dbl>
#>  1 death from cancer Drug A  0      0         0     
#>  2 death from cancer Drug A  3.53   0         0     
#>  3 death from cancer Drug A  5.33   0         0     
#>  4 death from cancer Drug A  6.32   0         0     
#>  5 death from cancer Drug A  7.27   0.0102    0.0102
#>  6 death from cancer Drug A  7.38   0.0204    0.0144
#>  7 death from cancer Drug A  8.37   0.0204    0.0144
#>  8 death from cancer Drug A  8.71   0.0204    0.0144
#>  9 death from cancer Drug A  9.24   0.0204    0.0144
#> 10 death from cancer Drug A  9.73   0.0204    0.0144
#> # … with 434 more rows

glance(cuminc)
#> # A tibble: 1 × 8
#>   outcome_1   statistic_1  df_1 p.value_1 outcome_2  statistic_2  df_2 p.value_2
#>   <chr>             <dbl> <dbl>     <dbl> <chr>            <dbl> <dbl>     <dbl>
#> 1 death from…        1.99     1     0.159 death oth…      0.0886     1     0.766