Competing Risks Regression

Competing Risks Regression

# S3 method for formula
crr(formula, data, failcode = NULL, ...)

crr(x, ...)

# S3 method for default
crr(x, ...)

Arguments

formula
formula with Surv() on LHS and covariates on RHS. The event status variable must be a factor, with the first level indicating 'censor' and subsequent levels the competing risks.

data
data frame

failcode
Indicates event of interest. If failcode= is NULL, the first competing event will be used as the event of interest. Default is NULL.

...
passed to methods

x
input object

Value

tidycrr object

See also

Examples

crr(Surv(ttdeath, death_cr) ~ age + grade, trial)
#> 11 cases omitted due to missing values
#> 
#> ── crr() ───────────────────────────────────────────────────────────────────────
#> • Call Surv(ttdeath, death_cr) ~ age + grade
#> • Failure type of interest "death from cancer"
#> 
#> Variable   HR     SE      95% CI       p-value    
#> age        1.01   0.010   0.99, 1.03   0.55       
#> gradeII    1.06   0.361   0.52, 2.16   0.87       
#> gradeIII   1.54   0.324   0.82, 2.90   0.18