Competing Risks Cumulative Incidence

Competing Risks Cumulative Incidence

# S3 method for formula
cuminc(formula, data, strata, rho = 0, ...)

cuminc(x, ...)

# S3 method for default
cuminc(x, ...)

Arguments

formula
formula with Surv() on LHS and covariates on RHS. The event status variable must be a factor, with the first level indicating 'censor' and subsequent levels the competing risks.

data
data frame

strata
stratification variable. Has no effect on estimates. Tests will be stratified on this variable. (all data in 1 stratum, if missing)

rho
Power of the weight function used in the tests.

...
passed to methods

x
input object

Value

tidycuminc object

Examples

# calculate risk for entire cohort -----------
cuminc(Surv(ttdeath, death_cr) ~ 1, trial)

#> 

#> ── cuminc() ────────────────────────────────────────────────────────────────────

#> 

#> • Failure type "death from cancer"

#> time   estimate   std.error    
#> 5.00   0.000      0.000        
#> 10.0   0.030      0.012        
#> 15.0   0.120      0.023        
#> 20.0   0.215      0.029        
#> 

#> • Failure type "death other causes"

#> time   estimate   std.error    
#> 5.00   0.005      0.005        
#> 10.0   0.025      0.011        
#> 15.0   0.090      0.020        
#> 20.0   0.205      0.029        

# calculate risk by treatment group ----------
cuminc(Surv(ttdeath, death_cr) ~ trt, trial)

#> 

#> ── cuminc() ────────────────────────────────────────────────────────────────────

#> 

#> • Failure type "death from cancer"

#> strata   time   estimate   std.error    
#> Drug A   5.00   0.000      0.000        
#> Drug A   10.0   0.020      0.014        
#> Drug A   15.0   0.071      0.026        
#> Drug A   20.0   0.173      0.039        
#> Drug B   5.00   0.000      0.000        
#> Drug B   10.0   0.039      0.019        
#> Drug B   15.0   0.167      0.037        
#> Drug B   20.0   0.255      0.043        
#> 

#> • Failure type "death other causes"

#> strata   time   estimate   std.error    
#> Drug A   5.00   0.010      0.010        
#> Drug A   10.0   0.020      0.014        
#> Drug A   15.0   0.082      0.028        
#> Drug A   20.0   0.204      0.041        
#> Drug B   5.00   0.000      0.000        
#> Drug B   10.0   0.029      0.017        
#> Drug B   15.0   0.098      0.030        
#> Drug B   20.0   0.206      0.040        
#> 

#> • Tests

#> outcome              statistic   df     p.value    
#> death from cancer    1.99        1.00   0.16       
#> death other causes   0.089       1.00   0.77