Broom methods for tidy cuminc objects

Broom methods for tidy cuminc objects

# S3 method for tidycuminc
tidy(x, times = NULL, conf.int = TRUE, conf.level = x$conf.level, ...)

# S3 method for tidycuminc
glance(x, ...)

Arguments

x

object of class 'tidycuminc'

times

Numeric vector of times to obtain risk estimates at

conf.int

Logical indicating whether or not to include a confidence interval in the tidied output. Defaults to FALSE.

conf.level

Level of the confidence interval. Default matches that in cuminc(conf.level=) (typically, 0.95)

...

not used

Value

a tibble

tidy() data frame

The returned tidy() data frame returns the following columns:

Column Name	Description

outcome	Competing Event Outcome

time	Numeric follow-up time

estimate	Risk estimate

std.error	Standard Error

n.risk	Number at risk at the specified time

n.event	Number of events at specified time

n.censor	Number of censored observations at specified time

cum.event	Cumulative number of events at specified time

cum.censor	Cumulative number of censored observations at specified time

If tidy(time=) is specified, then n.event and n.censor are the cumulative number of events/censored in the interval. For example, if tidy(time = c(0, 12, 18)) is passed, n.event and n.censor at time = 18 are the cumulative number of events/censored in the interval (12, 18].

p-values

The p-values reported in cuminc(), glance.tidycuminc() and add_p.tbl_cuminc() are Gray's test as described in Gray RJ (1988) A class of K-sample tests for comparing the cumulative incidence of a competing risk, Annals of Statistics, 16:1141-1154.

Confidence intervals

The confidence intervals for cumulative incidence estimates use the recommended method in Competing Risks: A Practical Perspective by Melania Pintilie.

$$x^{exp(±z * se / (x * log(x)))}$$

where \(x\) is the cumulative incidence estimate, \(se\) is the standard error estimate, and \(z\) is the z-score associated with the confidence level of the interval, e.g. \(z = 1.96\) for a 95% CI.

See also

Examples

cuminc <- cuminc(Surv(ttdeath, death_cr) ~ trt, trial)

tidy(cuminc)
#> # A tibble: 228 × 12
#>     time outcome     strata estimate std.error conf.low conf.high n.risk n.event
#>    <dbl> <chr>       <chr>     <dbl>     <dbl>    <dbl>     <dbl>  <int>   <int>
#>  1  0    death from… Drug A   0         0      NA         NA          98       0
#>  2  3.53 death from… Drug A   0         0      NA         NA          98       0
#>  3  7.27 death from… Drug A   0.0102    0.0102  8.84e-4    0.0503     97       1
#>  4  7.38 death from… Drug A   0.0204    0.0144  3.90e-3    0.0652     96       1
#>  5  8.37 death from… Drug A   0.0204    0.0144  3.90e-3    0.0652     95       0
#>  6 10.1  death from… Drug A   0.0204    0.0144  3.90e-3    0.0652     94       0
#>  7 10.3  death from… Drug A   0.0204    0.0144  3.90e-3    0.0652     93       0
#>  8 10.5  death from… Drug A   0.0204    0.0144  3.90e-3    0.0652     92       0
#>  9 11.4  death from… Drug A   0.0306    0.0175  8.17e-3    0.0798     91       1
#> 10 11.8  death from… Drug A   0.0306    0.0175  8.17e-3    0.0798     90       0
#> # … with 218 more rows, and 3 more variables: n.censor <int>, cum.event <int>,
#> #   cum.censor <int>

glance(cuminc)
#> # A tibble: 1 × 8
#>   outcome_1    statistic_1  df_1 p.value_1 outcome_2 statistic_2  df_2 p.value_2
#>   <chr>              <dbl> <dbl>     <dbl> <chr>           <dbl> <dbl>     <dbl>
#> 1 death from …        1.99     1     0.159 death ot…      0.0886     1     0.766